refactor(materializer): extract scene-text-limits (file #3)
Lift truncateDisplayText + applySceneTextLimits (and TRUNCATION_SUFFIX) into
scene-text-limits.ts; materializer re-exports applySceneTextLimits. Pure text
truncation, no canonical state. Net 74/74 green, check 0.
